claude-agent-store/requirements.txt
--extra-index-url https://erc.timetoact-group.at/
erc3>=1.0.4
anthropic>=0.39.0
python-dotenv>=1.0.0
